\name{SBFLMMBayesian-package}
\alias{SBFLMMBayesian-package}
\alias{SBFLMMBayesian}
\docType{package}
\title{
\packageTitle{SBFLMMBayesian}
}
\description{
\packageDescription{SBFLMMBayesian}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{SBFLMMBayesian}

Maintainer: \packageMaintainer{SBFLMMBayesian}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = SBFLMMBayesian)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{SBFLMMBayesian}}
%\section{Documentation Index}{\packageIndices{SBFLMMBayesian}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
